// src/types/master/ItemTypes.d.ts
export interface ItemTableProps {
  loading: boolean;
  handleEditData: (data: ItemTableData) => void;
  itemTableInput: string;
  handleFilterTableData: (value: string) => void;
  currentPage: number;
  setCurrentPage: Dispatch<SetStateAction<number>>;
  lastPage: number;
  perPage: number;
  onPerPageChange: (perPage: number) => void;
  showDeleteDialog: boolean;
  setShowDeleteDialog: Dispatch<SetStateAction<boolean>>;
  setTempDeleteId: Dispatch<SetStateAction<number | null>>;
  handleShowDeleteDialog: (id: number) => void;
  handleDeleteItem: () => void;
  deleteItemLoading: boolean;
  deleteWarning: string | null;
}

export interface ItemTableData {
  Id: number;
  Cat_Id: numberl;
  Model_Id: number | null;
  Size_Id: number | null;
  Color_Id: number | null;
  Item_Name: string;
  Item_Sh_Name: string;
  Unit_Id: number;
  Purchase_Gl: number;
  Sales_Gl: number;
  CGST: string;
  SGST: string;
  IGST: string;
  Pur_Rate: string;
  Sale_Rate: string;
  Cat_Name: string;
  Open_Qnty: string;
  Item_Rate: string;
}